+++ src/web/src/resource/CommentToggle.vue
@@ -1,7 +1,6 @@
 <script setup>
 import {inject} from "vue";
-import Btn from "../kit/Btn.vue";
-import Icon from "../kit/Icon.vue";
+import SideToggle from "./SideToggle.vue";
 
 const props = defineProps({resource: {type: Object, default: null}});
 const talk = inject("talk", null);
@@ -10,23 +9,6 @@ const system = () => Boolean(props.resource && props.resource.data && props.reso
 
 <template>
     <template v-if="talk && !system()">
-        <Btn small :class="['comment-toggle', {on: talk.talking.value}]" @click="talk.toggle()">
-            <Icon name="bubble" :size="12" />
-            Comments
-            <template v-if="talk.count.value">
-                <span class="comment-count">{{ talk.count.value }}</span>
-            </template>
-        </Btn>
+        <SideToggle mode="comments" icon="bubble" label="Comments" :count="talk.count.value" />
     </template>
 </template>
-
-<style scoped>
-.comment-count {
-    color: var(--text-3);
-}
-
-.comment-toggle.on {
-    border-color: var(--accent);
-    color: var(--accent-text);
-}
-</style>